// src/app/models/post.ts
export interface Post {
    id: number;
    title: string;
    plan: string;
    address: string;
    opdo: number;
    price: number;
    area: number;
    bedrooms: number;
    bathrooms: number;
    summary: number;
    published_at: string;
    built_year: number;
    build_status: string;
    author: string;
    localidad: any;
    images: any[];
}